config: Add CMake variables for the path of Mbed OS program and config file
`MBED_ROOT` and `MBED_CONFIG_FILE` are provided in the generated configuration
CMake module. `MBED_ROOT` is used by application programs to add the
subdirectory where the Mbed OS program is located. `MBED_CONFIG_FILE` is used
in the application project as well as in the Mbed OS program to access the
variables in the generated configuration file.